feat: 2fa login only when enabled + bug fixes
