+++ client/src/app/home-page/src/lib/home-page.module.ts
@@ -9,6 +9,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatListModule } from '@angular/material/list';
 import { MatDividerModule } from '@angular/material/divider';
+import { MatDialogModule } from "@angular/material/dialog";
 
 @NgModule({
     imports: [
@@ -18,8 +19,9 @@ import { MatDividerModule } from '@angular/material/divider';
       MatGridListModule,
       MatCardModule,
       MatButtonModule,
-      MatIconModule, 
-      MatListModule
+      MatIconModule,
+      MatListModule,
+      MatDialogModule
     ],
     declarations: [
       HomePageComponent,
