+++ client/src/app/home-page/src/lib/home-page/home-page.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { Trip } from '@client/shared-models';
 import { AuthService, TripFilter, TripService } from "@client/shared-services";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   templateUrl: './home-page.component.html',
@@ -11,10 +12,12 @@ export class HomePageComponent implements OnInit {
   authService: AuthService = inject(AuthService);
   private _tripService: TripService = inject(TripService);
   private _router: Router = inject(Router);
+  private _dialog: MatDialog = inject(MatDialog);
 
   private _trips: Trip[] = []
 
   ngOnInit(): void {
+    this._dialog.closeAll();
     const filter = <TripFilter>{ onGoing: true };
     this._tripService.getTrips(filter).subscribe(trips => this._trips = trips);
   }
@@ -25,13 +28,13 @@ export class HomePageComponent implements OnInit {
 
   tripClicked(trip: Trip) {
     const latestTripStation = trip.routeStations[trip.routeStations.length - 1];
-    this._router.navigate(['/game/random-train'], 
-    { 
-      queryParams: { 
-        tripId: trip.tripId, 
-        uicCode: latestTripStation.uicCode, 
+    this._router.navigate(['/game/random-train'],
+    {
+      queryParams: {
+        tripId: trip.tripId,
+        uicCode: latestTripStation.uicCode,
         location: latestTripStation.mediumName
-      } 
+      }
     });
   }
